Cambios a varios controladores
-Comenzada la implementacion del controlador de personaje

+++ AppGM/AppGMCore/Modelos/Personajes/ModeloPersonaje.cs
@@ -18,7 +18,7 @@ namespace AppGM.Core
         public string Nombre { get; set; }
 
         //Stats del personaje
-        public short Hp { get; set; }
+        public int Hp { get; set; }
         public ushort Str { get; set; }
         public ushort End { get; set; }
         public ushort Agi { get; set; }
